# docs/openapi/recovery-v1.yaml
openapi: 3.0.3
info:
  title: InvoNotify Recovery API
  version: 1.0.0
  description: |
    Tenant-scoped recovery APIs for the AI Revenue Recovery demo. All write
    endpoints require an authenticated merchant session and reject cross-origin
    state changes. Recovery action execution is simulation-only by default.
servers:
  - url: http://localhost:3001
security:
  - sessionCookie: []
paths:
...
  /api/v1/promises/{promiseId}/reconcile:
    get:
      summary: Fetch promise reminders and outcomes
      parameters: [{ $ref: '#/components/parameters/promiseId' }]
      responses: { '200': { description: Promise detail } }
    post:
      summary: Reconcile fulfilled, partial, or missed promise against confirmed payment
      parameters: [{ $ref: '#/components/parameters/promiseId' }]
      responses: { '200': { description: Reconciliation result } }
...
  /api/v1/recovery-actions/{actionId}/execute:
    post:
      summary: Execute a bounded action in safe simulation mode
      parameters:
        - $ref: '#/components/parameters/actionId'
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { mode: { type: string, enum: [SIMULATION] } } } } }
      responses: { '200': { description: Simulated execution result }, '409': { description: Approval, terminal, or idempotency conflict } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: authjs.session-token }
  parameters:
...
    promiseId: { name: promiseId, in: path, required: true, schema: { type: integer, minimum: 1 } }
    actionId: { name: actionId, in: path, required: true, schema: { type: integer, minimum: 1 } }